Add files via upload
In this project we need to implement the Face Recognition using Deep Face and Face Net. For this we have a large data-set and need to implement Triplet Loss and based on the Triplet Loss we need to make predictions.

Due to large size of zip file I am unable to upload the whole zip file(file folder). So I wanna to provide the link of Google Drive Folder.
Files are uploaded here :)
https://drive.google.com/drive/folders/1So3q8LtCQVeAppqB0imGd0vyLibJ-H4M